requirements.txt
confluent-kafka
faker
kafka-python
duckdb
pandas
minio
deltalake